requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aiohttp=3.9.5=pypi_0
aiosignal=1.3.1=pypi_0
altair=5.3.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=4.3.0=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
backoff=2.2.1=pypi_0
beautifulsoup4=4.12.3=pypi_0
blinker=1.8.2=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.3.11=h06a4308_0
cachetools=5.3.3=pypi_0
certifi=2024.2.2=pypi_0
cffi=1.16.0=pypi_0
chardet=5.2.0=pypi_0
charset-normalizer=3.3.2=pypi_0
chroma-hnswlib=0.7.3=pypi_0
chromadb=0.3.29=pypi_0
click=8.1.7=pypi_0
clickhouse-connect=0.7.8=pypi_0
coloredlogs=15.0.1=pypi_0
contourpy=1.2.1=pypi_0
cryptography=42.0.7=pypi_0
cycler=0.12.1=pypi_0
dataclasses-json=0.6.6=pypi_0
dataclasses-json-speakeasy=0.5.11=pypi_0
deprecated=1.2.14=pypi_0
distro=1.9.0=pypi_0
duckdb=0.10.2=pypi_0
effdet=0.4.1=pypi_0
emoji=2.12.1=pypi_0
exceptiongroup=1.2.1=pypi_0
fastapi=0.85.1=pypi_0
filelock=3.14.0=pypi_0
filetype=1.2.0=pypi_0
flatbuffers=24.3.25=pypi_0
fonttools=4.51.0=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.5.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
greenlet=3.0.3=pypi_0
h11=0.14.0=pypi_0
hnswlib=0.8.0=pypi_0
httpcore=1.0.5=pypi_0
httptools=0.6.1=pypi_0
httpx=0.27.0=pypi_0
huggingface-hub=0.21.3=pypi_0
humanfriendly=10.0=pypi_0
idna=3.7=pypi_0
importlib-metadata=7.1.0=pypi_0
iopath=0.1.10=pypi_0
jinja2=3.1.4=pypi_0
joblib=1.4.2=pypi_0
jsonpatch=1.33=pypi_0
jsonpath-python=1.0.6=pypi_0
jsonpointer=2.4=pypi_0
jsonschema=4.22.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
kiwisolver=1.4.5=pypi_0
langchain=0.1.9=pypi_0
langchain-community=0.0.24=pypi_0
langchain-core=0.1.35=pypi_0
langchain-openai=0.0.8=pypi_0
langchain-text-splitters=0.0.1=pypi_0
langchainhub=0.1.14=pypi_0
langdetect=1.0.9=pypi_0
langsmith=0.1.26=pypi_0
layoutparser=0.3.4=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
lxml=5.2.2=pypi_0
lz4=4.3.3=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
marshmallow=3.21.2=pypi_0
matplotlib=3.9.0=pypi_0
mdurl=0.1.2=pypi_0
monotonic=1.6=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.5=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.3=pypi_0
nltk=3.8.1=pypi_0
numpy=1.26.4=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.19.3=pypi_0
nvidia-nvjitlink-cu12=12.4.127=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
omegaconf=2.3.0=pypi_0
onnx=1.16.0=pypi_0
onnxruntime=1.15.1=pypi_0
openai=1.13.3=pypi_0
opencv-python=4.9.0.80=pypi_0
openssl=3.0.13=h7f8727e_2
orjson=3.10.3=pypi_0
overrides=7.7.0=pypi_0
packaging=23.2=pypi_0
pandas=2.2.1=pypi_0
pdf2image=1.17.0=pypi_0
pdfminer-six=20221105=pypi_0
pdfplumber=0.10.4=pypi_0
pikepdf=8.13.0=pypi_0
pillow=10.3.0=pypi_0
pillow-heif=0.15.0=pypi_0
pip=24.0=py310h06a4308_0
portalocker=2.8.2=pypi_0
posthog=3.5.0=pypi_0
protobuf=4.25.3=pypi_0
pulsar-client=3.5.0=pypi_0
pyarrow=15.0.0=pypi_0
pycocotools=2.0.7=pypi_0
pycparser=2.22=pypi_0
pydantic=1.10.14=pypi_0
pydeck=0.9.1=pypi_0
pygments=2.18.0=pypi_0
pyparsing=3.1.2=pypi_0
pypdf=4.0.2=pypi_0
pypdf2=3.0.1=pypi_0
pypdfium2=4.27.0=pypi_0
pytesseract=0.3.10=pypi_0
python=3.10.14=h955ad1f_1
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
python-iso639=2024.4.27=pypi_0
python-magic=0.4.27=pypi_0
python-multipart=0.0.9=pypi_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
rank-bm25=0.2.2=pypi_0
rapidfuzz=3.9.1=pypi_0
readline=8.2=h5eee18b_0
referencing=0.35.1=pypi_0
regex=2024.5.15=pypi_0
requests=2.32.0=pypi_0
rich=13.7.1=pypi_0
rpds-py=0.18.1=pypi_0
safetensors=0.4.2=pypi_0
scikit-learn=1.4.1.post1=pypi_0
scipy=1.12.0=pypi_0
sentence-transformers=2.4.0=pypi_0
sentencepiece=0.1.99=py310hdb19cb5_0
setuptools=69.5.1=py310h06a4308_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.5=pypi_0
sqlalchemy=2.0.27=pypi_0
sqlite=3.45.3=h5eee18b_0
starlette=0.20.4=pypi_0
streamlit=1.31.1=pypi_0
streamlit-option-menu=0.3.12=pypi_0
sympy=1.12=pypi_0
tabulate=0.9.0=pypi_0
tenacity=8.3.0=pypi_0
threadpoolctl=3.5.0=pypi_0
tiktoken=0.5.2=pypi_0
timm=1.0.3=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.15.2=pypi_0
toml=0.10.2=pypi_0
toolz=0.12.1=pypi_0
torch=2.2.1=pypi_0
torchvision=0.17.1=pypi_0
tornado=6.4=pypi_0
tqdm=4.66.4=pypi_0
transformers=4.38.2=pypi_0
triton=2.2.0=pypi_0
types-requests=2.31.0.20240406=pypi_0
typing-extensions=4.11.0=pypi_0
typing-inspect=0.9.0=pypi_0
tzdata=2024.1=pypi_0
tzlocal=5.2=pypi_0
unstructured=0.12.4=pypi_0
unstructured-client=0.18.0=pypi_0
unstructured-inference=0.7.23=pypi_0
unstructured-pytesseract=0.3.12=pypi_0
urllib3=2.2.1=pypi_0
uvicorn=0.29.0=pypi_0
uvloop=0.19.0=pypi_0
validators=0.28.1=pypi_0
watchdog=4.0.0=pypi_0
watchfiles=0.21.0=pypi_0
websockets=12.0=pypi_0
wheel=0.43.0=py310h06a4308_0
wikipedia=1.4.0=pypi_0
wrapt=1.16.0=pypi_0
xz=5.4.6=h5eee18b_1
yarl=1.9.4=pypi_0
zipp=3.18.2=pypi_0
zlib=1.2.13=h5eee18b_1
zstandard=0.22.0=pypi_0
